requirements.txt
Django==3.1
django-filter==2.4.0
djangorestframework==3.12.4
Markdown==3.3.4
django-cors-headers==3.8.0
djangorestframework-simplejwt==4.8.0
mixer==7.1.2
